REFACTOR: blitz-utils -> blitz-models
