+++ pyproject.toml
@@ -43,8 +43,8 @@ dev = [
 ]
 
 [project.urls]
-"Homepage" = "https://github.com/Jylpah/blitz-utils"
-"Bug Tracker" = "https://github.com/Jylpah/blitz-utils/issues"
+"Homepage" = "https://github.com/Jylpah/blitz-models"
+"Bug Tracker" = "https://github.com/Jylpah/blitz-models/issues"
 
 
 [build-system]
@@ -52,7 +52,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/blitzutils"]
+packages = ["src/blitzmodels"]
 
 [tool.hatch.metadata]
 allow-direct-references = true
