+++ src/blitzmodels/wotinspector/wi_apiv2.py
