+++ tests/test_map.py
@@ -14,7 +14,7 @@ debug = logger.debug
 
 sys.path.insert(0, str(Path(__file__).parent.parent.resolve() / "src"))
 
-from blitzutils import Map, Maps, MapMode, MapModeStr
+from blitzmodels import Map, Maps, MapMode, MapModeStr
 
 
 ########################################################
