+++ tests/test_release.py
@@ -13,7 +13,7 @@ debug = logger.debug
 sys.path.insert(0, str((Path(__file__).parent.parent / "src").resolve()))
 
 from pyutils import export, awrap
-from blitzutils import Release
+from blitzmodels import Release
 
 
 ########################################################
