+++ tests/test_tank.py
@@ -16,14 +16,14 @@ debug = logger.debug
 
 sys.path.insert(0, str(Path(__file__).parent.parent.resolve() / "src"))
 
-from blitzutils import (
+from blitzmodels import (
     Tank,
     EnumNation,
     EnumVehicleTier,
     EnumVehicleTypeInt,
     EnumVehicleTypeStr,
 )
-from blitzutils import WGApiWoTBlitzTankopedia
+from blitzmodels import WGApiWoTBlitzTankopedia
 
 
 ########################################################
